Make NativeCall::Types::CArray an Iterable
Fixes R#2681.  FWIW, the tests of the module pass ok now.  But I'm a bit
baffled by the fact that there is no visible "iterator" method in the class,
so I'm not 100% this is the solution.  It *is* however 'make test' and
'make spectest' clean, and it makes all of the tests of Crypt::TweetNacl
pass.