Add Buf/Blob.STORE candidates for non-iterables
For some reason we cannot make things like CArray an Iterable without
breaking a lot of things (like R#2687).  This fixes R#2681 the hard way.
If the native types are the same, this could be sped up quite a bit, but
I don't see a quick way to introspect that and fast-path that.