Introduce IO::Resolver
This class will be the superclass for all DNS resolvers. This contains a
resolve method, which is used to perform an OS-level DNS resolution, and
a lookup method, which performs a DNS lookup, but produces addresses
more suitable for connections. Both methods return a Supply of
IO::Address::Info. An instance of an IO::Resolver lives in $*RESOLVER,
and will be used by IO::Socket::INET and IO::Socket::Async to handle
resolving hostnames passed to their methods.