Add method and resolver named parameters to socket creation methods
The resolver named parameter contains an IO::Resolver instance to use
for DNS resolution; the method named parameter determines which method
of the resolver given to use to handle DNS resolution. This is done to
allow people to easily revert to v6.c DNS resolution behaviour for
connections in v6.e if it is desirable to do so.