Refactor IO::Socket::INET socket initialization
The !initialize method was copied from a Perl 5 library. This assumes
that only SOCK_STREAM sockets may be connected, but this is wrong;
SOCK_DGRAM sockets may be connected as well, and it is desireable for us
to support this. Without the SOCK_STREAM logic, this method can be
simplified and split into two methods: !CONNECT and !BIND.