Remove IO::Socket::INET host/port splitting in v6.e
There is a feature specific to IO::Socket::INET where hostnames would be
parsed as potential URIs, which would split a port from an IP literal if
one exists.  The current implementation of this is buggy; all URIs are
treated as IPv4 URIs when PF_UNSPEC is passed, breaking parsing of IPv6
literals (like ::1). While PF_UNSPEC URI parsing can be implemented, it
is rather complex to handle in a spec-compliant way (this would require
a full-blown grammar to accomplish, not just a couple regexen).

Because this behaviour is unique to IO::Socket::INET, doesn't work
properly, cannot work properly without complicating IO::Socket::INET
significantly, and doesn't provide much benefit in the first place, I
don't think this belongs in core.