Allow null hostnames in v6.e's IO::Socket::INET
The upcoming v6.e changes to DNS resolution pose a problem: the hosts
file is not respected, so localhost cannot be resolved! This is a common
hostname to use with connections, so we need to provide an alternative
to encourage people to use instead. While this is typically ::1 and
127.0.0.1, we cannot encourage people to use these addresses explicitly;
when PF_UNSPEC is the family given (which is the default), whether or
not localhost will resolve to ::1 or 127.0.0.1 depends on how the system
is configured, and how configuration is handled varies greatly from
platform to platform.

v6.c's IO::Resolver.resolve adds a feature to our existing DNS support:
support for null hostnames. These are resolved to the loopback address
for addresses intended for connections or the unspecified address for
bindings, respecting how the system is configured. This cannot be used
by v6.c's IO::Socket::INET since it already defaults to using 0.0.0.0 as
an address when none is given for servers, so it's handled in v6.e
instead.