Deprecate IO::Socket::INET localhost/localport/host/port methods
Besides these methods no longer being adequate for representing socket
address state, the host methods of IO::Socket::INET are not consistent
with IO::Socket::Async's: they return the host arguments given upon
socket initialization, not the IP literal for them. These are to be
removed in v6.e in favour of the local-address/remote-address methods.