Generically bind curried roles to their ultimate parameterization
This adds a new bind metamethod to Metamodel::CurriedRoleHOW. Backed by
synonymous metamethods on Metamodel::ParametricRoleGroupHOW or any other
HOW, this will eventually allow for us to fix edge cases in role
typechecking.

We need to update Metamodel::CurriedRoleHOW's parameterize_roles and
update_role_typecheck_list methods to depend on this immediately, as
having an extra attribute loafing around makes builds horrifically
slow. As a consequence, we wind up cutting time off builds instead. This
could imply that only doing this for role groups meant selecting a
candidate multiple times for each individual role, but I'm not sure on
the exact source of this.