Loosen archetype checks during curried role parameterization
Later on, we need roles added to curried roles to be composable or
composalizable. In further constraining this to generic and parametric
types, it becomes very difficult for a curried role to do anything other
than another role. Enums are an example of this from Rakudo, but this
alone isn't enough for that to work.