Invert role typechecks in Metamodel::ParametricRoleHOW
Deferring to accepts_type on doees here is incorrect. We can't guarantee
that a doee is instantiable at this point in time, so typechecks fail
when they shouldn't: "Cannot find method ACCEPTS on type T". That
old and tiresome error is gone now, but now we have role group problems.

While we're at it, decont. We can't guarantee there won't be a container
when HOWs we don't control enter the picture.